requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-classy-tags==4.0.0
django-cms==3.11.0
django-formtools==2.4
django-phonenumber-field==7.0.0
django-sekizai==4.0.0
django-treebeard==4.5.1
djangocms-admin-style==3.2.0
idna==3.4
phonenumberslite==8.13.1
Pillow==9.3.0
python-dotenv==0.21.0
pytz==2022.6
reportlab==3.6.12
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tinycss2==1.2.1
tzdata==2022.7
Unidecode==1.1.2
urllib3==1.26.13
webencodings==0.5.1
